Make API async
This makes `format` and `lint` async. This was needed to support the linters out there that does not have a sync API.

The LinterAdapter interface supports `formatAsync`, `formatSync`, `lintAsync` and `lintSync`. At least one format and lint function needs to be present on the linterAdapter that is being registered.

In the `format` and `lint` functions sync versions of the LinterAdapter's functions take presidence over the async versions. I don't know if this the best choice?